Add directional assert to Source/Sink
This adds to the Source and Sink components an assert on m_flow direction
similar to what is already in the PartialVolume components.

Some models in Examples such as SimpleStream and WaterHammer do
not have volumes so without this commit there is no check for invalid
mass flow direction if users modify the models.

With this commit, some topologies may still have invalid flow direction
without an assert. To make a complete check, the SISOFlow or Inlet
and Outlet connectors need an assert, but that will lead to many more
assert messages. This commit is a compromise to catch most errors
with fewer messages.